+++ sentry-spring/src/main/java/io/sentry/spring/SentryInitBeanPostProcessor.java
@@ -1,7 +1,9 @@
 package io.sentry.spring;
 
 import com.jakewharton.nopen.annotation.Open;
+import io.sentry.EventProcessor;
 import io.sentry.ITransportFactory;
+import io.sentry.Integration;
 import io.sentry.Sentry;
 import io.sentry.SentryOptions;
 import io.sentry.SentryOptions.TracesSamplerCallback;
@@ -18,6 +20,7 @@ public class SentryInitBeanPostProcessor implements BeanPostProcessor, Applicati
   private @Nullable ApplicationContext applicationContext;
 
   @Override
+  @SuppressWarnings("unchecked")
   public Object postProcessAfterInitialization(
       final @NotNull Object bean, @NotNull final String beanName) throws BeansException {
     if (bean instanceof SentryOptions) {
@@ -37,6 +40,23 @@ public class SentryInitBeanPostProcessor implements BeanPostProcessor, Applicati
         applicationContext
             .getBeanProvider(ITransportFactory.class)
             .ifAvailable(options::setTransportFactory);
+        applicationContext
+            .getBeanProvider(SentryOptions.BeforeSendCallback.class)
+            .ifAvailable(options::setBeforeSend);
+        applicationContext
+            .getBeanProvider(SentryOptions.BeforeBreadcrumbCallback.class)
+            .ifAvailable(options::setBeforeBreadcrumb);
+        applicationContext
+            .getBeansOfType(EventProcessor.class)
+            .values()
+            .forEach(options::addEventProcessor);
+        applicationContext
+            .getBeansOfType(Integration.class)
+            .values()
+            .forEach(options::addIntegration);
+        applicationContext
+            .getBeanProvider(Sentry.OptionsConfiguration.class)
+            .ifAvailable(optionsConfiguration -> optionsConfiguration.configure(options));
       }
       Sentry.init(options);
     }